Add support for sleep using timers

+++ include/sys/defs.h
@@ -22,6 +22,6 @@ typedef uint64_t off_t;
 
 typedef uint32_t pid_t;
 typedef uint32_t mode_t;
-typedef uint8_t bool;
+typedef uint32_t bool;
 
 #endif
